Added Pump <-> FuelTank interface.

+++ Assignment1/PumpStatus.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "..\Transaction.h"
 struct PumpStatus{
-	int quantityFueled;			// Progress in the fueling process
+	double quantityFueled;			// Progress in the fueling process
 	int fuelGrade;
 	bool complete;
 	bool pumpOn = true;				// Kill the pump (set by the GSC)
